+++ game_patch/misc/alpine_settings.cpp
@@ -1015,6 +1015,16 @@ bool alpine_player_settings_load(rf::Player* player)
         g_alpine_game_config.player_join_flash = std::stoi(settings["PlayerJoinFlash"]);
         processed_keys.insert("PlayerJoinFlash");
     }
+    if (settings.contains("PlayerJoinFlashHighlightOnly")) {
+        g_alpine_game_config.player_join_flash_highlight_only =
+            std::stoi(settings["PlayerJoinFlashHighlightOnly"]);
+        processed_keys.insert("PlayerJoinFlashHighlightOnly");
+    }
+    if (settings.contains("PlayerJoinFlashCancelable")) {
+        g_alpine_game_config.player_join_flash_cancelable =
+            std::stoi(settings["PlayerJoinFlashCancelable"]);
+        processed_keys.insert("PlayerJoinFlashCancelable");
+    }
     if (settings.contains("PlayerJoinFlashTimeoutSeconds")) {
         g_alpine_game_config.player_join_flash_timeout_sec =
             std::stoul(settings["PlayerJoinFlashTimeoutSeconds"]);
@@ -1490,6 +1500,8 @@ void alpine_player_settings_save(rf::Player* player)
     file << "DisableMultiCharacterLOD=" << g_alpine_game_config.multi_no_character_lod << "\n";
     file << "PlayerJoinBeep=" << g_alpine_game_config.player_join_beep << "\n";
     file << "PlayerJoinFlash=" << g_alpine_game_config.player_join_flash << "\n";
+    file << "PlayerJoinFlashHightlightOnly=" << g_alpine_game_config.player_join_flash_highlight_only << "\n";
+    file << "PlayerJoinFlashCancelable=" << g_alpine_game_config.player_join_flash_cancelable << "\n";
     file << "PlayerJoinFlashTimeoutSeconds=" << g_alpine_game_config.player_join_flash_timeout_sec << "\n";
     file << "WorldHUDAltDamageIndicators=" << g_alpine_game_config.world_hud_alt_damage_indicators << "\n";
     file << "DesiredHandicap=" << g_alpine_game_config.desired_handicap << "\n";