+++ game_patch/misc/alpine_settings.h
@@ -163,6 +163,8 @@ struct AlpineGameSettings
     int seasonal_effect = 1; // 0=none, 1=auto, 2=always_snow
     bool player_join_beep = false;
     bool player_join_flash = true;
+    bool player_join_flash_highlight_only = false;
+    bool player_join_flash_cancelable = false;
     uint32_t player_join_flash_timeout_sec = 0;
     bool full_range_lighting = true;
     bool always_clamp_official_lightmaps = false;